Moved seeds data to dedicated folder Created seeding utility to avoid code-repetition
